Networking WIP

+++ GameEngine/Font.cpp
@@ -95,7 +95,7 @@ void Font::LoadFont(const wchar_t* filename, int windowWidth, int windowHeight)
 
 	Font font;
 	std::wstring tmp;
-	int startpos;
+	size_t startpos;
 
 	// extract font name
 	fs >> tmp >> tmp; // info face="Capture It"
